+++ packages/@webex/internal-plugin-mercury/test/unit/spec/socket.js
@@ -39,7 +39,7 @@ describe('plugin-mercury', () => {
       clock.uninstall();
     });
 
-    beforeEach('mock WebSocket and open a Socket', () => {
+    beforeEach(() => {
       sinon.stub(Socket, 'getWebSocketConstructor').callsFake(
         () =>
           function (...args) {