+++ packages/@webex/plugin-meetings/test/unit/spec/reachability/request.js
@@ -4,6 +4,7 @@ import MockWebex from '@webex/test-helper-mock-webex';
 import Meetings from '@webex/plugin-meetings';
 import ReachabilityRequest from '@webex/plugin-meetings/src/reachability/request';
 import {IP_VERSION} from '@webex/plugin-meetings/src/constants';
+import {NewMetrics} from '@webex/internal-plugin-metrics';
 
 
 describe('plugin-meetings/reachability', () => {
@@ -14,6 +15,7 @@ describe('plugin-meetings/reachability', () => {
     webex = new MockWebex({
       children: {
         meetings: Meetings,
+        newMetrics: NewMetrics
       },
     });
 
@@ -22,19 +24,25 @@ describe('plugin-meetings/reachability', () => {
       regionCode: 'WEST-COAST',
     };
 
-    webex.internal = {
-      services: {
-        get: sinon.mock().returns('locusUrl'),
-        waitForCatalog: sinon.mock().returns(Promise.resolve({})),
-      },
+    webex.internal.services = {
+      get: sinon.mock().returns('locusUrl'),
+      waitForCatalog: sinon.mock().returns(Promise.resolve({})),
     };
-
 
     reachabilityRequest = new ReachabilityRequest(webex);
 
   });
 
   describe('#getClusters', () => {
+
+    beforeEach(() => {
+      sinon.spy(webex.internal.newMetrics.callDiagnosticLatencies, 'measureLatency');
+    });
+
+    afterEach(() => {
+      sinon.restore();
+    });
+
     it('sends a GET request with the correct params', async () => {
       webex.request = sinon.mock().returns(Promise.resolve({
         body: {
@@ -49,7 +57,6 @@ describe('plugin-meetings/reachability', () => {
       }));
 
       const res = await reachabilityRequest.getClusters(IP_VERSION.only_ipv4);
-
       const requestParams = webex.request.getCall(0).args[0];
 
       assert.equal(requestParams.method, 'GET');
@@ -63,6 +70,7 @@ describe('plugin-meetings/reachability', () => {
       });
       assert.deepEqual(res.clusters.clusterId, {udp: "testUDP", isVideoMesh: true})
       assert.deepEqual(res.joinCookie, {anycastEntryPoint: "aws-eu-west-1"})
+      assert.calledOnceWithExactly(webex.internal.newMetrics.callDiagnosticLatencies.measureLatency, sinon.match.func, 'internal.get.cluster.time');
     });
   });
 });